remove Debian/Ubuntu maintainer references
The control.in file had historic references to Debian and Ubuntu teams and
individuals responsible for packaging gnome-shell which are not relevant to
Cinnamon.

There's now no reason to process control.in to generate control anymore, so
this file could probably be renamed control. I haven't done this in this
commit because the generate control file is in source control and appears
to have been modified.