Implementation of 'gravity padding' for applets:
Add an extra setting in the gconf listing for applet positions -
a percentage value of the screen width.  A slider for this becomes
available when panel edit mode is activated, in the context menu of
each applet.  The user can slide the value (the full range of the
slider theoretically represents the width of the screen) and we apply
a 'buffer' object with that much 'padding' to the 'pull' or gravity side
of the applet.

This has the net effect of allowing you to position applets where you
want them along the panel, by creating padding in between applets
in a given panel location (left, right, middle) - note the padding is
reversed in the right zone, since gravity is reversed.

Features:

- Slider appears in context menu of applet during panel-edit-mode
- Reset padding option is available in the panel settings->
  troubleshooting submenu
- Padding resets when an applet is dragged to a new location