Change the panel zone allocation to simulate 'centering' items in the center zone, rather than left-justified.  This makes panel-edit-mode not quite as pretty, but works for centering items up.
Unfortunately, the underlying boxlayout object and its parents don't
allow for a whole lot of control regarding the alignment and direction
of items.  We could x_align to middle but that requires us to turn on
expand as well - which would fill out the pre-allocated size and space
out the items unintentionally.

If we contrain the center zone and do the x_align to middle, that forces
the right zone to be larger, making apparent the fact that all items work
from left to right, with no way (that I can see) to change them.  So we're
forced to keep the right zone small to simulate right-justified objects,
and the center zone just constantly adjusts its left edge to half the
width of the screen minus half of the zone's width.