Fix for pidgin systray icon not appearing until a Cinnamon restart.
From what I can tell, when Pidgin is first started up, it's providing
an icon of 1px x 1px in size (using get_size on the icon object passed
from the dispatcher).  We can resize the height, or the width of this icon,
but not BOTH - it doesn't like that for some reason.  I've set it to
use height for now, as that seems to work best for all the apps I tried.