Clean up move workspace code and fix inconsistencies (#548)
With this code, it no longer wraps around when switching workspaces: i.e. no longer go to the last workspace, if you go to the left from the first one. Ctrl-Shift-Alt-Left/Right never did this.
If we want to introduce wrap around for everything, we should patch meta_workspace_get_neighbor() in muffin/src/core/workspace.c instead of doing it in an inconsistent Cinnamon-only way: Ctrl-Shift-Alt-Left/Right is handled in Muffin.