Add dependencies to settings widgets -
For instance, if you have a checkbox A, and you want combobox B and C
to be disabled if checkbox A is unchecked, you would pass the
schema/key combination for checkbox A along with the other construction
arguments for B and C.  If there is no dependency, pass None in its place.

Currently only supporting checkbox dependencies - anything more complicated
and it needs its own handling.

You can also invert the dependency by adding a ! in front of the
whole schema/key string.  In the above example, if inverted, if checkbox A is
UNCHECKED, combobox B and C WILL be enabled, and disabled if checkbox A
is CHECKED.

Argument structure:
[!]<dependency-schema>/<dependency-key>