Add invisible auto-scrolling regions to the top and bottom of the application list.
Each region has 3 individual sub-regions - slow, medium, fast.

Hovering the mouse in any of these regions scrolls the list up or down
at slow, medium or fast speed.

A small amount of application list real-estate is lost, but I think it's
a fair tradeoff to gain smooth scrolling.  The drawback to the old method
was that you had to keep the mouse moving to generate enter/leave events
on the application buttons, otherwise the scrolling would stop.