Fix for notification dead area (and probably all occurrences of dead area)
We were depending on _freezeUpdateCount to be 0 or false before running
updateRegion() in layout.js.  The value always ends up at -1 after startup
because we called freeze once, and thaw twice (for the 2 panelboxes).

This whole concept could be rethought a bit - but for now, let's not allow
the freeze count to dip below 0 and affect updateRegion.