Window tracker: Strip extension from window class when using it for app lookup.
The WM_CLASS is set to the app full name (i.e. cinnamon-settings.py) and then
it uses that to try to find the app during certain operations.  It appends .desktop
to that string, and searches for the desktop entry, and, of course, does not find it.

app-system: check settings hash table for app as well as the normal table, when doing
a search.

All of this goes towards addressing the problem with cinnamon and mint applications
showing up as pixelated in the alt-tab view.