Update files/usr/lib/cinnamon-settings/cinnamon-settings.py
Added the option to show the menu icon in the title bar as well, since many metacity themes support it, (Clearlooks even uses the fancy logo of the application which is running inside the window, which I like) and people should be able to enable it, if there's an option to edit the title bar anyway. Otherwise people might be led to think that there's no option to enable this feature.