Implement 'advanced mode' for individual settings in the python modules.
Advanced mode can be toggled at any time and the current panel will be
re-calculated to show the appropriate settings.

The individual widget is called advanced by passing a True along with the
widget in sidePage.add_widget(widget, True) - if it is a normal option
the 2nd argument may be omitted.