Add standalone wallpaper changer..
It seems to work fine except for the fact that I get 'failed to create drawable'
messages, which doesn't sound good.  I wonder if the bg object is expiring before
the X server gets around to actually changing the pixmap, but I get the warning
even if I add time.sleep(5) or something.