Add option to modify the alt-tab switcher delay
The alt-tab switcher uses a default delay of 150 ms before actually
showing the dialog to avoid repeated flashing. The downside of this
approach is that it makes it appear to be laggy. This commit adds the
ability to modify the delay of the dialog in the Window settings with
150 ms remaining the default choice to maintain the previous behavior
for people who prefer it.

Signed-off-by: Niklas Koep <niklas.koep@gmail.com>