st-label.c: Prevent .xsession-errors overflow - a common error/oversight when using timers that update a label occasionally, is forgetting to disconnect these timers when the widgets they act on go away or are replaced by a duplicate new object.
When this happens, you can get a cumulative increase in log spam as these
timers add up, and continue to try to act on bad objects.

This doesn't do anything to prevent this bad behavior, but it will squelch
the log messages to one time only per StLabel instance, and provide a bit
of debugging information so these errors can be fixed.