Update applet.js
Changed the flow on which you change between windows using the mouse scroll. At first I thought it was just me, but hey, this is inconsistent with the workspace switcher, for example. And also other programs like Google Chrome(changing between tabs with the scroll).

(This is my first contribution to a project, so sorry if I was inconvenient in any way)