menu applet: Don't reset the category when dragging an app.
ClutterText sends a "text-changed" signal when it loses focus
(like when we start a DnD operation) - if the search box is entry
when this happens, the applet was taking this to mean the user just
cleared the search box, and clear the current category.

It will now skip this if a) a search wasn't previously active, and
b) the search box is empty.