Fix DST related weeknumber calculation errors
In locales with DST diffDate is missing 1 hour which leads to dayNumber
being one off. This leads to wrong week numbers for the summer time
period, if Jan 1st is a Thursday.

For CET there CW 13 was double every few years, e.g. 2009 and 2015, and
CW 43 is missing.

Using 12 noon instead of midnight for the next Thursday date works
around this, since the 1h DST difference has no influence and gets
floored either way in the dayNumber calculation.