Add extension versioning support
This commit provides the option for extensions to ship different
versions for different Cinnamon verions. To enable this, extensions have
to set

   "multiversion": true

in metadata.json. The different versions should be put in different
subdirectories. For example, the version for 2.6 should be put in
<extension-uuid>/2.6/. The contents in 2.6/ should be exactly what you
will normally put in <extension-uuid>/, apart from metadata.json, which
should always be put in the parent directory, ie. <extension-uuid>/.

You do not need to create a subdirectory for every single Cinnamon
version out there. Cinnamon looks for the most recent subdirectory that
is not newer than the running Cinnamon version. For example, if you are
running 2.6 and there are 2.4, 2.8 directories, the 2.4 directory will
be loaded, until the you upgrade to 2.8 in which case the 2.8 directory
will be used. Minor version numbers can also be used, eg. 2.6.4, and are
sorted accordingly.

If no suitable directory is found, then the contents in
<extension-uuid>/ will be loaded. Note that Cinnamon versions prior to
2.6 will not understand this directory magic, and will always try to
load the contents in <extension-uuid>/. Hence it is suggested that
extension maintainers put the version for 2.4 in <extension-uuid>/, and
create new directories if changes _specific to newer cinnamon versions_
are made. Don't make a new directory whenever a Cinnamon version is out
since it is just a waste of space (and maintenance effort).

Note that here "extensions" refers to what is normally known as
extensions, applets and desklets (but not themes). All these extensions
can use this versioning scheme to support multiple versions.