Use a sane way to find extension directories.
In the past, if we wanted to load the applet applet@cinnamon.org, we go
to /usr/share/cinnamon/applets/ (and other similar directories), list
everything in the directories, and for each item we check if it is a
directory, and see if the name is the same as the applet uuid. This is
stupid.

A much saner way of doing so would be to load
/usr/share/cinnamon/applets/applet@cinnamon.org/ directly.