Fix panel launcher dnd
Panel Launcher DND used to rely on the application name to identify
panel launchers. This causes unpredictable behaviour during dnd when two
applications of the same name are in the panel launcher (easily
reproduced by placing the same application twice). This fixes the issue
by directly identifying the launchers by the actual launcher object.

This fixes #3474 (the issue mentioned by the thread-hijacker. The
original issue was already fixed).