Support keybindings to toggle horizontal/vertical maximized state
Some graphical applications are ment to be controlled mainly though the
keyboard, think terminal windows.  For these it's extremely handy to have
keybindings to toggle horizontal or vertical maximized state, so one
doesn't have to reach out to the mouse.

Support setting keybindings for 'toggle-maximized-horizontally' and
'toggle-maximized-vertically' in Cinnamon Settings.