[Alt-F2] Redo (broken) completion mechanism
The old code fails to complete filesystem paths properly, eg. looking
for files in Desktop/stuff/more_stuff/. Also the code is complex, with
vastly different code doing the completion for the command and the
arguments.

This commit rewrites the completion mechanism, borrowing some from the
old code and some from GNOME Shell js/ui/runDialog.js, and removes a lot
of code duplication. It also improves the scrolling when there is a
large number of possible completions (in the old mechanism, the list
will go off-screen, and now it is ellipsized).