[Alt-F2] Fix completion for symlink to directory
For some reason, the Gio thinks that a symlink to a directory is not a
directory, even though it is asked to follow symlinks and return
properties of the target. This is a problem since we automatically
append a "/" if the path completed is a directory, while we add a " "
when it is a regular object.

The solution is, apparently, to tell Gio that we want to know the file
type (in addition to the name) when creating the FileInfo object. This
is mysterious since without this fix, Gio is still able to provide a
corret file type unless it is a symlink.