Don't register Cinnamon with cinnamon-session
Cinnamon became a registered app with the following commit:
   https://github.com/linuxmint/Cinnamon/commit/a5dd0778078fa0ccdfdf7056a5a0299586dbe357

However, the following commit now makes cinnamon-settings-daemon survive cinnamon-session "Stop" calls:
    https://github.com/linuxmint/cinnamon-settings-daemon/commit/6c4edd6b4e244fc4d37ae93a2192421186c9aff0

This means CSD maintains the appearance of mapped windows to the end of the session and unregistered apps no longer
look weird.

For these unregistered to retain their titlebar, the WM needs to be alive as well. Making Cinnamon an unregistered app
solves that issue.