xlet settings: Replace old binding function in settings.js with two new ones
The primary reason for this is that the binding direction is no longer meaningful internally due to the use of javascript properties rather than using a watch function. For that reason, the 'bindProperty' function is now considered deprecated.

In its place are two new functions:
bind works the same as bindProperty, but without the need to pass a binding direction as the first argument.
bindWithObject works identically to bind but allows the caller to specify the object to which the property will be bound.

The bindProperty function still works as a wrapper around bind, but it is recommended that bind or bindWithObject be used in newer code.

The system applets and desklets were updated to reflect this change.