keyboard layout applet: Fix showing the correct flags.
gnome-kbd is buggy...

.get_short_group_names() returns names that don't match the flag names.
For instance, "en" for US English

.get_group_name(group) returns the correct short group name that
matches the flag name of the country in question.  This returns
"us" (correct) to find the flag name.

get_group_names() seems fine.

I had Arabic installed, and was getting the flag for Argentina -
Arabic is "ara", and Argentina is "ar".