applet.js: Properly layout applets in vertical panels
Currently applets do not expand and center properly in vertical panels. Fix this
and add a simple convenience function to hide the label in TextIcon type
applets.

This breaks the layout of many of our stock text icon applets. This is
temporary and will be fixed in follow up commits.