added PopupSwitchIconMenuItem and used it in sound-applet
The new PopupSwitchIconMenuItem is the same as PopupSwitchMenuItem, but allows icons to the left of the label.
(comparable with PopupMenuItem and PopupIconMenuItem)

It is used in the contextMenu of the sound-applet, to show icons for mute switches.
This commit is a response to the feature-request https://github.com/linuxmint/Cinnamon/issues/4940,
but it only implements the requested icons. It is missing the requested alignment of the text of 'Applications' and 'Output devices'.