panel-launchers: Redesign the launcher context menu
Now that we have support for launcher specific actions in the menu, put some of
the less used options in a submenu. Also add an additional submenu to hold the
menu items common to the right click context menu for most applets. This also
gives an easy way to remove a single instance of the panel launchers. Something
that is currently missing.