[Xlet settings] Added option to "colorchooser" elements to enable alpha.
By default, "colorchooser" elements can't choose a color with alpha
(transparency). Contrary to what it occurs in the current stable Cinnamon
version 3.0.7, where all "colorchooser" elements from xlet's settings windows
are able to select a color with alpha.

This commit adds to the "colorchooser" elements a new boolean option (optional)
called "enable-alpha". When set to **true**, "colorchooser" elements will be
able to select a color with alpha. When set to **false** or ignored,
"colorchooser" elements will not be able to select a color with alpha.

This commit closes linuxmint/Cinnamon#5647