menu-applet: Make the 'menu-background' style class work again
The new popupMenu implementation allows theming the menus based on orientation.
This overwrites the menu-background style class added by the menu applet. To
get around this connect to the actors 'style-changed' signal so that anytime
popupMenu.js changes the style class we ensure the applets custom style class
gets re-added.