[applet.js] Added a mechanism to highlight applets with open menus.
Since the removal of box pointers (menu arrows) from Cinnamon's UI,
there isn't a visual guide that inform us to which applet the currently
opened menu belongs to.

This commit adds a mechanism that will set or unset the "checked" pseudo class
to applets with menus. When the main menu of an applet (or its context menu)
is open, the "checked" pseudo class will be applied to the applet. When either
of the previously mentioned menus are closed, the "checked" pseudo class will
be removed.

Only *"standard"* applets are affected. *"Special"* applets like
**Window list**, **Panel launchers** and **Workspace switcher** aren't
affected and might need to be *"dealt with"* individually.

Cinnamon themes need to be modified to style the .applet-box:checked selector.

Removed trailing spaces and added missing semicolons.