panel-launchers: close the menu before destroying when removing launcher
If we destroy the menu while it is open then a callback set by pushModal removes
the grab when it recieves the actor's destroy signal. Shortly after, we attempt to
remove it again in popupMenu.js via the menu state change signal and fail. Closing
the menu before destroying the actor avoids this issue.

Fixes #5968