cinnamon-json-makepot: Added support to scan for JS files inside sub-dirs.
Currently the cinnamon-json-makepot command, when run with the --js argument,
it scans only the JavaScript files that are in the root folder of an xlet.
In an xlet with "multiversion" enabled, all JavaScript files inside
sub-directories will be ignored and the resultant .pot file will be incomplete.

This commit adds the ability to scan all JavaScript files inside an xlet folder
by saving a list with all JavaScript files found into a temporary file. Then that file is used to pass the
--files-from argument to xgettext.