sound applet: Add rhythmbox to the list of known players
This ensures Rhythmbox is present as a launchable player on a fresh
install when Rhythmbox is part of the default software selection.

Without this, Rhythmbox would be detected over MPRIS by the applet
and added to the list on its first run.

This fixes an integration paper cut so we don't have to wait for this to happen.