Calendar applet: Allow custom date formats in vertical panels
Not a lot will fit, but it's not up to us to decide what does and what doesn't.
The user starts with a sane %H%n%M default and a panel width which just about
fits that much.

It's up to him or her then to decide to enlarge the panel and to use custom
formats. We shouldn't stop him/her from being able to do that.