cinnamon-settings: Do a better job of handling long labels
We are running into cases where either translations or long label lengths in
applet settings are putting a horizontal scrollbar on the window. This can cause
the "action" widget, like a switch, to be hidden outside of the viewable area.
To better deal with these situations, line wrap the labels instead. This ensures
that whole settings widget is viewable when opening the window at the default
size.