Layout: Ignore NaN numbers when updating regions
In Cinnamon 3.2, the "Window List with App Groups" applet
caused issues. The panel and windows wouldn't be clickable
anymore.

This was caused by NaN being returned by
let [x, y] = actorData.actor.get_transformed_position() and
let [w, h] = actorData.actor.get_transformed_size()

and an exception being thrown when doing:

let rect = new Meta.Rectangle({ x: x, y: y, width: w, height: h});

It's ok for the applet to fail if it references obsolete code
but it shoudn't affect Cinnamon as a whole.

When numbers are NaN, skip the actor altogether to ensure the regions
are being updated correctly.

Ideally we would log the issue, but this loop is too sensitive and
loops way too fast for that, it would make Cinnamon lag significantly
to do that.