JS API: Add Promise and setTimeout polyfills to environment.js
This adds a polyfill for Promise using the MIT licensed library from
https://github.com/taylorhakes/promise-polyfill. setTimeout is also
polyfilled because the Promise polyfill needs it. Adding it to the
global context will also help with compatibility for other third party
libraries. It is a wrapper around Mainloop.timeout_add.

On line 299, an extra return was added to avoid the "anonymous function
does not always return a value" CJS warning.

On line 9, an extra semicolon was removed.

This was tested with my own applets and confirmed working.