window-list: Don't lose our focus style class when opening the context menu
We don't lose the focus styling for a window list item when opening the right
click menu for an already focused window. While technically I suppose the menu
now has the focus instead of the window, that isn't what we want for styling
purposes. So ensure that the 'focus' style class isn't lost in this case.

Closes: https://github.com/linuxmint/Cinnamon/issues/5862