panel.js: move clipping code into new _setClipRegion helper
Clipping code is moved out of show and hide functions into a separate helper.
We now track the actor's shadow box by grabbing it in _moveResizePanel. This
is called on style changes, so this should update any time the box changes.
This will allow us to set the correct clipping in _moveResizePanel when it is
refactored in a later commit.