[WIP] load xlet after version check anyway
E.g. you use Cinnamon 3.2
If you want to add an xlet, which has the cinnamon-version flag in it's metadata.json file, but Cinnamon 3.2 is not included in this.
You get the following message:
> Extension %s is not compatible with current version of cinnamon. Using it may break your system. Load anyway?
> Yes, No

If you press yes, the xlet won't load anyway.

Allow xlet to load, if clicked Yes. Don't throw an error, log warning into LookingGlas instead.