window-list-applet: fix undefined reference warning
We are checking if the AppMenuButton is temporary, but the temporary property
is actually stored as 'alert'. This corrects that typo and in addition changes
all existing calls to _addWindow() to include the alert argument for consistency.
This ensures that window.alert is always initialized to either true or false.

fixes js38 warning at line 1233:
reference to undefined property window.temp