User applet: Remove ability to switch to guest with LightDM
LightDM doesn't necessarily provide guest sessions.

The user applet would need to parse files in /usr/share/lightdm and /etc/lightdm/lightdm.conf.d/
to assess whether or not this option is available before showing it.

Unless/until this is done, the applet cannot assume the availability of guest sessions
and shouldn't therefore show an action for it.

Note that users can still "switch users" and open a guest session from the DM itself.