photoframe-desklet: Fix a few issues
The xlet settings widget used to choose a folder for the images was changed to
use a URI instead of path. The code in the applet was never updated to reflect
that so choosing a new directory was broken.

Clean up a few of the signal values while as well.