hotCorner.js: Fix hover delay callback.
Fixes: https://bugs.launchpad.net/linuxmint/+bug/1702482

Zeroing out the delay id was occurring immediately after setting it,
so none of the cancel opportunities could ever run, resulting in the
hotcorner always getting activated once it was triggered.

Cleaned up some of the returns in our other event handlers as well,
to more clearly show their behavior.