popupMenu.js: cancel tweens on open/close if we are already animating
This prevents a race condition that can occur if a menu is opened and closed
in rapid succession. I don't believe this is reproducible by a user event, but
was exposed by a case where we had 2 menu activations, one immediately after
the other. The first menu would be closed by a grab loss signal during or
before the open animation started. When this happens the menu becomes stuck
open and unreactive as it has no grab.

The open tween is added on idle sometime after open() returns, so I'm not
sure of the exact order of events that cause the close animation to fail
and the menu to become stuck open with no grab.