window-list: set AppMenuButtons unreactive during panel edit mode
This prevents multiple menus from opening at once.

In this case, during panel edit mode, the event is first handled by the
AppMenuButton. This handler opens the item context menu but does not consume
the event and it is propagated down to the applet actor which also opens a
menu. This only happens during edit mode as the applet actor is only set
reactive during that time.