workspace.js: use our normal prototype init method
and avoid using arguments when unnecessary. This fixes zoom on scale
view. I'm not sure of the exact cause of this error, but my best guess
is that setting an array from arguments directly is not supported. MDN
defines arguments as an Array-like object and proposes using
Array.from(arguments) when you need it to be an array. Since we have
static named arguments here we don't need to bother messing with arguments.

Fixes: #6707

JS ERROR: TypeError: arguments[Symbol.iterator] is not a function
ScaledPoint@/usr/share/cinnamon/js/ui/workspace.js:65:5