layout.js: special case handling of panel actors
Instead of using the actor's clip region to determine the input region,
special case hidden panel actors and constrain the input region to the
geometry of the monitor it is configured for.

This continues to prevent panels from causing input region on adjacent
monitors when autohide is enabled while avoiding the errors introduced
by the clip region method.

This also fixes monitor detection for panel actors. If we detect the
actor's monitor by its location on screen, then we end up not hiding the
small exposed strip of a hidden panel while the true monitor is
inFullscreen.