layout.js: special case handling of panel actors
This adds two special cases for handling of hideable panel actors.

The first special case is in Chrome.getMonitorInfoForActor() this
function returns a monitor by finding where the actor's rect places
on the screen. Since panel actors are moved off monitor and cliped,
this can be incorrect in certain multimon configurations. This reads
the configured monitorIndex from the panel actor's delegate and
returns the corresponding monitor object.

The second case is a replacement for the current clip region behavior
which affects any actor with a clip region set. It causes issues when
scaling is involved, when the actor is clipped larger than its
allocation, and general rounding errors on translation to stage
coordinates. Instead of relying on the clip, we now just intersect
the transformed actor rect with the monitor rect to find the input
region. This still constrains the input region to one monitor, but
now only affects hideable panel actors.

In both of these cases it is possible for the Panel.monitorIndex to
be invalid. This will occur during refreshes after a monitor
reconfiguration before the panel object has handled reconfiguration.

Functions calling getMonitorInfoForActor() are expected to handle
null results, and should handle undefined just as well. In
updateRegions() the monitor object is checked, and in the worst case
we add a 0 size input rect to the input region when the panel doesn't
intersect with the monitor we got.