windows-quick-list: wrap menu in scrollable submenu container
This wraps the menu items in a PopupSubMenuMenuItem menu.
PopupSubMenuMenuItem was modified so its label and Clutter.Point can be
skipped in its _init if no label text is passed to it. The classes for
PopupSubMenuMenuItem's actor and menu actor are reset in the applet's
_init method, so the applet retains its proper theming.

Closes #2491

Check the submenu for its open state before toggling it, avoid allowing a null property to be assigned to an actor object because the binding doesn't like that