layout.js: optimize signals a bit
Moves signals that only touch chrome into chrome's prototype. Some
of these signals do make more sense in LayoutManager based on what
they react to, but in practice there is no value added by the extra
abstraction as we aren't doing anything with them in LayoutManager.
The duplicate restacked signal hook is eliminated. The behavior of
Chrome._windowsRestacked is modified to match the behavior of the
two signals.