panel.js: refactor updatePanelVisibility()
Moves the intelligent autohide code out into its own function, uses
ifs instead of a switch, and avoids checking at all when in panel
edit mode because we always force a show in that case.

This makes the code quite a bit easier to read for the simple
cases, and will make it easier to add additional always-show
conditions that short circuit the autohide logic.