panel.js: "peek" panel for 1.5 seconds if hidden while highlighted
This causes a panel to show briefly if it is hidden while highlight
is turned on. This is done by panel settings and applet settings to
indicate which panel you are working on. With hidden panels, it can
be very difficult to notice as only a 1 pixel strip is visible.