Restructure appletManager state, replace uuidMap, idMap, and extension._loadedDefinitions with an array of objects, reduce the amount of promise functions used
We were creating three references to an object containing references to
other objects, in order to look them up by UUID, applet_id, or to sort
the definitions by UUID. This consolidates them to a single collection -
an array of objects that can be looked up with the repurposed
getAppletDefinition function. The old getAppletDefinition function was
renamed to createAppletDefinition to better indicate what its doing. In
addition, createAppletDefinition checks if its creating a repeat object
before doing so by checking the array, this will prevent increasing the
complexity and redundancy of the resulting references to panel, the
location, and the applet instances.

Iterating arrays is much faster than iterating objects in JS, so in
addition to these cyclical reference checks, it should also improve
performance.

Update mapping in cinnamonDBus