Update cs_power.py
Allow for longer times till the system goes to sleep - added 2 and 3 hours. 
This is useful when listening to albums ( > 60 min) without having to move the mouse or watching movies. 
Maybe this could/should be fixed else where for all different media/video players and maybe browsers (html5 or flash) but given the simplicity of this solution I guess this is useful already - See feature request  #6185

I tested the "fix" (actually just additional options) but the translation(s) for "hours" are missing afterwards. 
Where or how can I add translations?