Fix issues related to Cinnamon running with cjs/mozjs52
This fixes some of the mozjs52 related warnings occurring in the logs
except for extension related modules, which will be handled in #6878.

This also fixes a few string.replace calls so they only use regex, the
old way of only passing strings to `replace` was not on a standards
track and will break in JS.

Some unused imports were also cleaned up.