Don't filter "JavaEmbeddedFrame" windows
Although we don't want these windows to show up in our window list,
it's really not up to us to fix this.

Our solution consisted in checking the window title and though added cost
to the WM on each map signal.

Java can make this window skip_taskbar, override_redirect or even git it an app,
assuming it needs the window in the first place.