Window list: Make AppButtons listen to their own window directly
Previous to this commit, the applet was listening globally for window change
from the tracker and the window manager. It then looped on the appbuttons
to find the one corresponding to the right metawindow.

Not only that but on focus change, every single appbutton was asked to
check its focus state.

This commit simplifies this. The applet no longer listens to global events.
Instead, each appbutton reacts to its own metawindow signals.