Speed up windows mapping (#7251)
* Don't filter "JavaEmbeddedFrame" windows

Although we don't want these windows to show up in our window list,
it's really not up to us to fix this.

Our solution consisted in checking the window title and though added cost
to the WM on each map signal.

Java can make this window skip_taskbar, override_redirect or even git it an app,
assuming it needs the window in the first place.

* Window list: Make AppButtons listen to their own window directly

Previous to this commit, the applet was listening globally for window change
from the tracker and the window manager. It then looped on the appbuttons
to find the one corresponding to the right metawindow.

Not only that but on focus change, every single appbutton was asked to
check its focus state.

This commit simplifies this. The applet no longer listens to global events.
Instead, each appbutton reacts to its own metawindow signals.

* Window list: Create the window buttons asynchronously

Delay the creation of the window buttons by 20ms. Using idle_add results in
delaying way too much. 20ms is not noticeable by the user and it no longer delays
the mapping of the window itself.

This gives us performance close to metacity on window creations.

* Window list: Include the _shouldAdd() call into the timeout as well.

* Bring back the fix for JavaEmbeddedFrame windows

Now that we're building things asynchronously, this doesn't cost much anymore.

* Panel launchers: Only animate the launcher after launching the app

Also remove some unecessary allocation maths.