Support for environment colors in themes
Colors can be defined like if they were "variables" and picked by the user
at runtime. This is done via a CSS function (similar to rgb) called "env".
It can be used as env("name") where name is one of the supported
"variables", inspired by the getenv function in C. This can be used in
borders, backgrounds or anywhere a color is supported.

This only supports rgb colors, because typically the theme wants to play
with the alpha channel rather than always using a fully opaque color. So the
"env" function has an optional second parameter where you can define the
alpha like in rgba. For example: env("name", 0.5)

This behavior can be disabled by the user (and it's disabled by default), so
that one can use the theme defaults if they want. To define a default color
the theme just needs to make use of the fallback feature of CSS.

I have defined two "variables": accent and text. "Accent" is the color that
stands out in the theme. Also exists "text" so that one can pick a font
color that is visible against the accent color if it's used as a
background color.