Skip unnecessary panel loading in change handlers, fix settings destruction on desklet move
If an xlet is added, it probably won't be loaded by the time the signals
fire the change handlers - and since the xlets add themselves at the end
of their load routine - triggered by initEnabled at the end of the
function, there is no point in calling addAppletToPanels/_loadDesklet.

Further distinguished between removed/changed applets - in
deskletManager the change handler is firing twice, and `_loadDesklet` is
true first, then false again - likely because of the async timing being
off from the signal firing.