appletManager.js: fix getRunningInstancesForUuid()
Since we removed the discrete instances object appletObj and merged
instances into the definition object, we now have null instances.
This occurs during startup, for definitions that are on removed panels,
and likely for applets that fail to load.

To return the previous behavior we just need to filter instances that
aren't actually running at the moment, aka null.