notifications: close button and no fade on mouseover (#7336)
* messageTray.js: notification close button

Allows notifications to be individually dismissed before
or after entering the message tray.

This should return to the source that the notification was
dismissed, and does not activate the default action.

* notifications: remove fade on mouseover support

This was turned on by default for years and makes the close button harder
to use. If you need to see through the notification you can now quickly
dismiss it with the close button without activating the associated action.