Fix resident notifications not displaying (#7248)
* Fix resident notifications not displaying

processNotification() used to call pushNotification() if the
notification was resident and the app had focus.  This CL changes it
to always call notify() instead.

The only difference is that notify() additionally emits the 'notify'
signal which hooks into _onNotify(), the function that actually puts
the notification into the queue.  Without this, the notification would
never get displayed.

* Remove unused var