update applet.js
related to this feature request. 
https://github.com/linuxmint/Cinnamon/issues/6136
I change the code for this applet just to remove the "switch user" option when the gsettings key "org.cinnamon.desktop.lockdown" "disable-user-switching" is true. I need it in Linux Mint 18.3 so I change it in this version. 
Sorry for my english and maybe the code could be "cleaner" ?