notificationsDaemon: bring back clear on focus + blacklist
After using cinnamon without the clearing for a while, we've come to
the conclusion that we do need this behavior for excessively spammy
notification sources, at least until the notifications are reworked
so that it's possible to: 1) bypass/silence a huge queue of unimportant
or unwanted notifications, and 2) clear those notifications from the
tray en masse while preserving possibly useful notifications.

This commit brings back auto clear on focus and adds a blacklist.
Currently uses a hardcoded list, but this could be in gsettings instead
if we need user configurability. The intial list is just a few popular
browsers.

This reverts commit 65c4cc1e0469b32e685c3029dc6b1df2d45e9a09.