Notifications fixups (#7413)
* notifications: improve URLHighlighter performance

This is called on every hover event. Before, we iterated over every
position in the text. This was acceptable for shorter texts, but as
they grow in length performance drops dramatically.

* notifications: don't allow mouse scrolling

We currently have scrolling disabled for notifications, so we can
just disable mouse scrolling which fixes scrolling inside the message
tray (by not intercepting the scroll events inside the notification).

fixes: #6068