network applet: refresh the panel icon more frequently when the menu is open.
Signal strength for individual AP's, including the connected one, are
updated as the underlying device reports those changes.  The panel icon,
however, is on a fixed update schedule - updating only every 10 seconds
except when a connection is in the process of being made.

There's nothing wrong here really, except when you have the applet menu
open, the active connection's strength can be at odds with the icon in the
panel.  This will cause updates to the panel icon to happen every 2 seconds
when the menu is open, and revert to every 10 seconds when it's not.

Refactor additionally to prevent a new GSource being created during each
periodic update.