network applet: show the actual active connection's signal strength
in the AP list, rather than the strongest AP for a given ssid, which
may not be the same thing.

Remove the periodic timer for the panel icon and text updates.  Reacting
to a connection's strength changes directly results in an average of
around 10.5s between updates (measured over about 1 hour,) which is
slightly better than our periodic update, and has the added advantages
of our signal strength always being immediately updated when the underlying
property changes, and not having the additional overhead and logic of an
extra timer.

Note:
this is a fast libnm migration of the corrispective commit in master