chrome: Improve visibility conditions for panels and other chrome (#7552)
when windows are fullscreen and popups or menus are activated.

Intended behavior:

Single monitor setups: Panels are hidden only when a given fullscreen
window for a given monitor actually is the toplevel window for that
monitor.  If a popup is activate (like the Cinnamon menu via shortcut,)
the panels will show while that modal is active.  This prevents a
weird disembodied menu popping out of nowhere.  Further, if a non-
fullscreen window is alt-tabbed to, or otherwise made the focus, the
panels will also return until the fullscreen window is re-focused.

Multi-monitor setups: Same as single monitor, except for modal actor
behavior is different - only the panel(s) on the monitor which currently
hosts the last-added/current modal actor are shown.  This allows, for
instance, the alt-tab popup or menu to be used on your 'work' monitor
without disrupting the view of your 'movie/tv-show' monitor.

Requires muffin patch as well: https://github.com/linuxmint/muffin/pull/317