Revert "network applet: Don't show unmanaged networks in the list."
This reverts commit 3e366ce3f5cfe118ae1ab594bf2ce8c7872d8b4d.

This commit resulted in wireless devices not showing up in the applet
even though they were manageable and visible in CCC.

Although the commit properly hid unmanageable devices, it also stopped
the applet from listening to signal changes on them.

If a device was unmanageable when added and then manageable then-after
the applet wouldn't hear the signal change and the device wouldn't
show up in the menu.

This is reproducible on a Macbook Pro with:

sudo rmmod wl
sudo modprobe wl

When the driver comes back, the menu should rebuild and list the wireless APs.
This commit prevented that.