cinnamon-schema scripts: harden against dangerous input
These scripts are supposed to be run via pkexec as root. The current
implementation is very naive by trusting input parameters and passing
them to cp and rm via a shell subprocess. This allows wildcards to be
interpreted or path components to be added to the schema file to remove.
This can lead to the scripts performing way different operations than
originally intended.

With this change a couple of attack vectors are addressed:

- removal of files outside of /usr/share/glib-2.0/schemas is prevented
- reading from special devices that could cause the script to block
infinitely or to cause disk space exhaustion in the target directory is
prevented
- copying of files not owned by the user running pkexec is denied to
prevent information leaks
- copying of overly large input files is prevented