windowManager: Improve workspace button response time, introduce an API for calling callbacks after the next stage relayout
This adds a new API, afterStageRelayout to Main.wm. It is a
self-contained signal handler that connects to the stage's
"queue-relayout" signal with connect_after. This allows us to retain the
non-blocking workspace switching behavior without relying on Mainloop -
which isn't guarunteed to be in sync with Clutter.

Add missing semicolon

Remove doc formatting