window list applet: Don't display windows with skip-taskbar set.
I kind of think this should be addressed in Main.isWindowInteresting and
the check removed from CinnamonWindowTracker - in the main function, we
check the tracker, but only return there if the return is TRUE.  Dialog
windows with skip-taskbar set (which is common) end up being shown,
because they're not app-backed windows. The skip-taskbar hint should be
respected regardless.

Removing the check from the window tracker function would require changes
to not only windowAttentionHandler but also any xlets that make use of
either the main or the window tracker functions.  And since other entities
such as alt-tab and the quick list applet already do their own skip-taskbar
check, just fixing the window list is ok.