Make sure window lists reload when different instances are added or removed
In order to differentiate when other instances are changing we need to
pass those other instances to on_applet_instances_changed so the applet
can compare. Ideally this should be changed so its only called if they
are a different instance, but not sure if that behavior change would
break other applets.