Refresh only windows on instances changed
refreshCurrentAppList (aka AppList.prototype.refreshList) destroys all
actors in an AppGroup list, and recreates them. This is a heavy call, so
for places where only windows need to be reloaded, we can just reload
those instead.

This also wraps refreshList calls in a timeout because the delay can
affect the responsiveness of menus and settings toggles when many apps
are pinned.