gwl: Support dragging of unpinned apps
This hardens the pinning logic to ensure we always get the correct
favorites order by basing it on AppList's actor children, and filtering
by favorite from that order.

It also corrects a bug in acceptDrop where sometimes the icon could not
be placed right of its current position because GWL was fetching a
reference of the actor which may have been stale by that point, instead
of using `source.actor`.