Menu: Place system buttons beside the search field
This commit partly fixes the broken layout of the menu on the master branch,
after the merge of new-panel-layout.

Note the following remaining issues:

- New widgets (St.BoxLayout in particular) were added in the menu without
using any style name. These should be styled to allow themes to set padding.

- There should be some spacing by default in topRow between the search field
and the system buttons.

- When hovering an application with the mouse (it doesn't happen when selecting
it with the keyboard), the size of the search field slightly expands and that
results in the system buttons moving right.