Menu applet: Fix height
We don't want the menu to be too small.. 400px works well in many res
(except 800px).

We don't want it to big larger than the workarea..

And ideally it should be configurable to some extent.

This commit makes the min size of the favbox configurable, gives it
a good default of 400px which works everywhere and caps it by taking
the monitor height into consideration.

Two issues remain:

1. the user is asked to configure something which isn't intuitive...
the size of the faxbox, as opposed to the resulting size of the
entire menu.

2. the maths in this commit use an approximation when doing
(sysbuttonsbox.height * 2) to estimate the height of the components in the
menu which aren't the favbox itself.