gwl: Give the number label badge styling, misc. clean up (#7986)
* gwl: Give the number label badge styling

* gwl: Fix skip-taskbar windows appearing in the window list

* gwl: Fix the number label clipping the badge when more than 9 windows in a group are open

* gwl: Listen for icon changes from MetaWindow

* gwl: Improve vertical number label alignment relative to badge

* gwl: Account for borders when determining badge size

* gwl: Fix thumbnails clipping out of the monitor region

* gwl: Remove badge border work-around

* gwl: Fix invalid object access in refreshThumbnail

* gwl: Fix thumbnails not refreshing on instance change

This moves the initial loading of the app list to
on_applet_instances_changed. This combined with app list deferral helps
ensure the initial load can filter windows based on a complete monitor
watch list with all instances loaded and accounted for.

* gwl: Fix move to monitor skipping some windows

* gwl: Fix up thumbnail scaling re-work

Make sure the thumbnail refresh is always called after thumbnail
addition/removal

* Fix CodeFactor warning

* gwl: Ensure the thumbnail menu can toggle between the vertical and normal orientations when working around fitting many thumbnails

* gwl: Preserve focus on hover

* gwl: Clean up pseudo class handling

* gwl: Don't assume the window will close after closing from a thumbnail

* gwl: Default to 100% opacity for the hover peek effect

* gwl: Remove the closed pseudo class

Not needed.

* gwl: Merge item-box and thumbnail-box classes in grouped-window-list-thumbnail-menu

* gwl: Move thumbnail close button styling to constants

This moves the scaling offset value to state in applet.js, so it only
needs to be checked once, and when the scale changes. This prevents the
style string concatenation and offset value ternary being computed on
each thumbnail construction.

* gwl: Enforce a minimum size for thumbnails when scaling to fit

* gwl: Make sure the focus handler returns early while the thumbnail menu is open

Closes #8017

* gwl: Consider padding and margin in thumbnail scaling logic

Fixes thumbnail scaling in third party themes.

It also adds a helper function to St.Widget to help cut down on both the
boilerplate, and to make these methods less tedious to use.

* gwl: Don't center thumbnail titles by altering container width

This idea was first shown in @germanfr's PR (#8008), and is less
fragile.